+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
+
 - **New**
   - `Mirror Implement` Thaumaturge support
+- **Updated**
+  - Added aaa default pick for `Dancing Blade`
+  - Added a way in specific summons to exclude default summoned traits from some summon
 
 ## 2.5.4